HOP-12 Rename package
