--- plugins/databases/mysql/src/main/resources/org/apache/hop/core/database/messages/messages_en_US.properties
@@ -1,3 +1,0 @@
-MySQLDatabaseMeta.Exception.LegacyColumnNameNoDBMetaDataException="Please provide a valid DatabaseMetaData object"
-MySQLDatabaseMeta.Exception.LegacyColumnNameNoRSMetaDataException="Please provide a valid ResultSetMetaData object"
-MySQLDatabaseMeta.Exception.LegacyColumnNameException=Something unexpected went wrong trying to get the legacy column name
