+++ plugins/databases/mysql/src/test/java/org/apache/hop/database/mysql/MySQLValueMetaBaseTest.java
@@ -1,4 +1,4 @@
-package org.apache.hop.database;
+package org.apache.hop.database.mysql;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -30,7 +30,7 @@ import org.apache.hop.core.row.value.ValueMetaBase;
 import org.apache.hop.core.row.value.ValueMetaFactory;
 import org.apache.hop.core.row.value.ValueMetaPluginType;
 import org.apache.hop.core.row.value.ValueMetaString;
-import org.apache.hop.database.MySQLDatabaseMeta;
+import org.apache.hop.database.mysql.MySQLDatabaseMeta;
 import org.apache.hop.junit.rules.RestoreHopEnvironment;
 import org.junit.Before;
 import org.junit.BeforeClass;
